+++ 99-integration/dubbo-samples-mybatis/README.md
@@ -1,6 +1,5 @@
 # Dubbo + Mybatis
 
-This is a dubbo + mybatis sample.
+This is a dubbo + spring boot + mybatis sample.
 
-In this sample, provider is using @ComponentScan and consumer is using xml configuration.
