--- 99-integration/dubbo-samples-mybatis/src/main/java/org/apache/dubbo/samples/boundary/mybatis/impl/MybatisServiceImpl.java
@@ -1,48 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package org.apache.dubbo.samples.boundary.mybatis.impl;
-
-import org.apache.dubbo.config.annotation.DubboService;
-import org.apache.dubbo.samples.boundary.mybatis.api.MybatisService;
-import org.apache.dubbo.samples.boundary.mybatis.api.User;
-import org.apache.dubbo.samples.boundary.mybatis.impl.dao.UserDao;
-import org.apache.dubbo.samples.boundary.mybatis.impl.dao.UserModel;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-@DubboService(version = "1.0.0")
-public class MybatisServiceImpl implements MybatisService {
-
-    @Autowired
-    private UserDao userDao;
-
-    @Override
-    public User findByUserId(long id) {
-        UserModel userModel = userDao.findByUserId(id);
-
-        User u = new User();
-
-        u.setId(userModel.getId());
-        u.setName(userModel.getName());
-
-        return u;
-    }
-}
