--- 99-integration/dubbo-samples-mybatis/src/main/resources/spring/application-provider.properties
@@ -1,12 +1,0 @@
-spring.liquibase.change-log=classpath:liquibase/liquibase-changelog.xml
-spring.application.name=mybatis-provider
-dubbo.scan.base-packages=org.apache.dubbo.samples.boundary.mybatis
-dubbo.protocol.name=dubbo
-dubbo.protocol.port=20880
-dubbo.registry.address=zookeeper://127.0.0.1:2181
-dubbo.provider.token=true
-dubbo.application.name=${spring.application.name}
-
-spring.datasource.type=com.alibaba.druid.pool.DruidDataSource
-spring.datasource.driver=org.h2.Driver
-spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;MODE=Oracle;TRACE_LEVEL_SYSTEM_OUT=2
